+++ src/OFS/tests/testHistory.py
@@ -4,10 +4,10 @@ Zope2.startup()
 
 import os
 import shutil
+import time
 import transaction
 import tempfile
 import ZODB
-
 
 from OFS.Application import Application
 from OFS.History import Historical
@@ -42,10 +42,12 @@ class HistoryTests(unittest.TestCase):
         t.description = None
         t.note('Change 1')
         t.commit()
+        time.sleep(0.02) # wait at least one Windows clock tick
         hi.title = 'Second title'
         t = transaction.get()
         t.note('Change 2')
         t.commit()
+        time.sleep(0.02) # wait at least one Windows clock tick
         hi.title = 'Third title'
         t = transaction.get()
         t.note('Change 3')